Refactory controller code
