+++ TODOLIST.md
@@ -37,9 +37,9 @@
   - [x] 屏幕旋转: 这个指强制当前屏幕旋转至哪个方向
     - [ ] iPad 无效,暂不知原因
   - [x] 截图
-  - [ ]  悬浮窗中播放
-    - [ ] 悬浮窗的UI控制器
-    - [ ] 自定义UI
+  - [ ] 悬浮窗中播放
+    - [ ] 悬浮窗的 UI 控制器
+    - [ ] 自定义 UI
 - [x] 默认控制器 UI
   - [x] 进度条
   - [x] 播放/暂停按钮
@@ -74,6 +74,8 @@
   - [x] 在列表中(ListView)
   - [x] 视频竖向分页滑动
   - [x] 在悬浮窗中播放
-    - [x] 悬浮窗的UI控制器
-      - [ ] 允许自定义UI
+    - [x] 悬浮窗的 UI 控制器
 - [x] iOS 部分视频无法显示图像的问题: 可能很长时间内都无法解决
+- [ ] 支持在 dart 端初始化 ijkPlayer 播放器的 option
+  - [ ] android
+  - [ ] iOS
